chore: refactor app list into common component (#9754)
